Cleaned up OspreySharp assembly naming and enum file layout
* Dropped "pwiz." prefix from AssemblyName in all 8 .csproj (matches SkylineBatch/AutoQC convention; RootNamespace and C# namespaces unchanged)
* Updated InternalsVisibleTo attributes to track new assembly names
* Bumped AssemblyCopyright year from 2025 to 2026 in all 8 AssemblyInfo.cs
* Merged 7 simple enum .cs files into most-related class: DecoyMethod/FdrLevel/FdrMethod/ResolutionMode/SharedPeptideMode into OspreyConfig.cs, ToleranceUnit into FragmentToleranceConfig.cs, IonType plus IonTypeExtensions into FragmentAnnotation.cs
* All 186 unit tests pass, zero warnings

See ai/todos/active/TODO-20260409_osprey_sharp.md
