--- pwiz_tools/OspreySharp/OspreySharp.Core/FdrMethod.cs
@@ -1,36 +1,0 @@
-/*
- * Original author: Brendan MacLean <brendanx .at. uw.edu>,
- *                  MacCoss Lab, Department of Genome Sciences, UW
- * AI assistance: Claude Code (Claude Opus 4) <noreply .at. anthropic.com>
- *
- * Based on osprey (https://github.com/MacCossLab/osprey)
- *   by Michael J. MacCoss, MacCoss Lab, Department of Genome Sciences, UW
- *
- * Copyright 2026 University of Washington - Seattle, WA
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-namespace pwiz.OspreySharp.Core
-{
-    /// <summary>
-    /// Statistical method for FDR estimation.
-    /// Maps to osprey-core/src/types.rs FdrMethod.
-    /// </summary>
-    public enum FdrMethod
-    {
-        Percolator,
-        Mokapot,
-        Simple
-    }
-}
