+++ pwiz_tools/OspreySharp/OspreySharp.IO/OspreySharp.IO.csproj
@@ -9,7 +9,7 @@
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>pwiz.OspreySharp.IO</RootNamespace>
-    <AssemblyName>pwiz.OspreySharp.IO</AssemblyName>
+    <AssemblyName>OspreySharp.IO</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
