+++ pwiz_tools/OspreySharp/OspreySharp.ML/OspreySharp.ML.csproj
@@ -9,7 +9,7 @@
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>pwiz.OspreySharp.ML</RootNamespace>
-    <AssemblyName>pwiz.OspreySharp.ML</AssemblyName>
+    <AssemblyName>OspreySharp.ML</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
