+++ pwiz_tools/OspreySharp/OspreySharp.Scoring/OspreySharp.Scoring.csproj
@@ -9,7 +9,7 @@
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>pwiz.OspreySharp.Scoring</RootNamespace>
-    <AssemblyName>pwiz.OspreySharp.Scoring</AssemblyName>
+    <AssemblyName>OspreySharp.Scoring</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
