+++ pwiz_tools/OspreySharp/OspreySharp.Scoring/Properties/AssemblyInfo.cs
@@ -9,7 +9,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("University of Washington")]
 [assembly: AssemblyProduct("OspreySharp.Scoring")]
-[assembly: AssemblyCopyright("Copyright (c) University of Washington 2025")]
+[assembly: AssemblyCopyright("Copyright (c) University of Washington 2026")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
