+++ pwiz_tools/OspreySharp/OspreySharp.Test/OspreySharp.Test.csproj
@@ -10,7 +10,7 @@
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>pwiz.OspreySharp.Test</RootNamespace>
-    <AssemblyName>pwiz.OspreySharp.Test</AssemblyName>
+    <AssemblyName>OspreySharp.Test</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
