+++ pwiz_tools/OspreySharp/OspreySharp/OspreySharp.csproj
@@ -9,7 +9,7 @@
     <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>pwiz.OspreySharp</RootNamespace>
-    <AssemblyName>pwiz.OspreySharp</AssemblyName>
+    <AssemblyName>OspreySharp</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <TargetFrameworkProfile />
